==> dea_analysis/build_panel/1_reshaping_data_for_panel.R <==
library(xts)

## function to reshape data for panel -----
panelify <- function(x, value_name){
    for (i in 1:dim(x)[2]){
        assign("tmp", data.frame(cbind(gsub("-", "", index(x[, i])), gsub("ID_", "", names(x[, i])), as.vector(x[, i]))))

        names(tmp) <- c("quarter", "IDRSSD", value_name)

    assign(paste0("tmp", i),
           tmp, envir = panelify_env)
    }
}


## change for whatever data we are reshaping ------

## T1 L Ratio-----

data <- readRDS("../../1_querying_data_and_analysis/analyses/panel_data_analysis/get_panel_data/pulling_all_firms/data_capital_ratios/t1_leverage_ratio.rds")

### Reshaping data: x is the output from query (xts with IDRSSD in column name) ----

panelify_env <- new.env()

panelify(data, "t1_leverage_ratio")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_t1_leverage_ratio.rds")

## T1 RBCR Ratio-----

data <- readRDS("../../1_querying_data_and_analysis/analyses/panel_data_analysis/get_panel_data/pulling_all_firms/data_capital_ratios/t1_risk_based_capital_ratio.rds")

### Reshaping data: x is the output from query (xts with IDRSSD in column name) ----

panelify_env <- new.env()

panelify(data, "t1_RBCR_ratio")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_t1_RBCR_ratio.rds")



## Domestic Deposits-----

data <- readRDS("domestic_deposits.rds")

### Reshaping data: x is the output from query (xts with IDRSSD in column name) ----

panelify_env <- new.env()

panelify(data, "domestic_deposits")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_domestic_deposits.rds")


## Interest Expense-----

data <- readRDS("./total_interest_expense.rds")

### Reshaping data: x is the output from query (xts with IDRSSD in column name) ----

panelify_env <- new.env()

panelify(data, "total_interest_expense")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_total_interest_expense.rds")







## Total Assets----

data <- readRDS("../../1_querying_data_and_analysis/analyses/panel_data_analysis/get_panel_data/pulling_all_firms/data_BS/total_assets.rds")

### Reshaping data: x is the output from query (xts with IDRSSD in column name) ----

panelify_env <- new.env()

panelify(data, "total_assets")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_total_assets.rds")


## NPA 30 to 89 --------

data <- readRDS("../../1_querying_data_and_analysis/analyses/panel_data_analysis/get_panel_data/pulling_all_firms/non_performing_loans/npa_30_89.rds")

panelify_env <- new.env()

panelify(data, "npa_30_89")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_npa_30_89.rds")

## NPA 90 plus --------

data <- readRDS("../../1_querying_data_and_analysis/analyses/panel_data_analysis/get_panel_data/pulling_all_firms/non_performing_loans/npa_90_plus.rds")

panelify_env <- new.env()

panelify(data, "npa_90_plus")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_npa_90_plus.rds")

## NPA nonacc --------

data <- readRDS("../../../get_panel_data/pulling_all_firms/non_performing_loans/npa_nonacc.rds")

panelify_env <- new.env()

panelify(data, "npa_nonacc")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_npa_nonacc.rds")

## Net income --------

data <- readRDS("../../../get_panel_data/pulling_all_firms/data_IS/net_income.rds")

panelify_env <- new.env()

panelify(data, "net_income")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_net_income.rds")


## Total SB Loan Data Lagged! --------


## Total SB Loan Data Delt Lagged 1 Year --------

data <- readRDS("../../../get_panel_data/pulling_all_firms/data_small_business_loans/totSBloans_Delt_lagged_1_year.rds")

panelify_env <- new.env()

panelify(data, "totSBloans_Delt_lagged_1_year")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_totSBloans_Delt_lagged_1_year.rds")

## Total Number of SB Loan Data Delt  --------

data <- readRDS("../../../get_panel_data/pulling_all_firms/data_small_business_loans/num_loans/totNumSBloans_Delt.rds")

panelify_env <- new.env()

panelify(data, "totNumSBloans_Delt")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_totNumSBloans_Delt.rds")

## Total Number of SB Loan Data Delt Lagged 1  --------

data <- readRDS("../../../get_panel_data/pulling_all_firms/data_small_business_loans/num_loans/totNumSBloans_Delt.rds")

data <- lag(data, k = 4)

panelify_env <- new.env()

panelify(data, "totNumSBloans_Delt_lagged_1")

panelified <- do.call(rbind, as.list(panelify_env))

saveRDS(panelified, "panel_totNumSBloans_Delt_lagged_1.rds")
